feature/adds date range picker and search UI

+++ modules/data/lib/dtos/stock_data_dto.dart
@@ -9,7 +9,7 @@ class StockDataDto {
   double? low;
   double? last;
   double? close;
-  int? volume;
+  double? volume;
   String? date;
   String? symbol;
   String? exchange;
