+++ modules/domain/lib/models/stock_data.dart
@@ -4,7 +4,7 @@ class StockData{
   double? low;
   double? last;
   double? close;
-  int? volume;
+  double? volume;
   String? date;
   String? symbol;
   String? exchange;
